skip destroy fix

+++ spec/spec_helper.rb
@@ -38,7 +38,7 @@ RSpec.configure do |c|
     c.ssh.close if c.ssh
     options = Net::SSH::Config.for(c.host)
     if(!ENV['LOCAL'])
-	run("vagrant destroy #{c.host} -f")
+	run("vagrant destroy #{c.host} -f") unless ENV['SKIP_DESTROY']
 	WATCH.reset
 	run("vagrant up #{c.host} --provider=#{PROVIDER}")
 	run("vagrant provision #{c.host}")
